% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-mlr-paper.R
\name{benchmark_custom_no_models}
\alias{benchmark_custom_no_models}
\title{mlr::benchmark() wrapper}
\usage{
benchmark_custom_no_models(learner, task)
}
\arguments{
\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}

\item{task}{mlr task}

\item{resampling}{mlr resampling}
}
\description{
mlr::benchmark() wrapper
}
