% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fun-data-preprocessing.R
\name{boxcox_response}
\alias{boxcox_response}
\title{boxcox_response}
\usage{
boxcox_response(data, response)
}
\arguments{
\item{data}{(\code{data.frame})\cr data.frame.}

\item{response}{(\code{character})\cr Name of response.}
}
\value{
\code{data.frame}
}
\description{
Performs a boxcox transformation on the response variable.
}
