% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fun-hyperspectral.R
\name{calc_nri_indices}
\alias{calc_nri_indices}
\title{Calculate Normalized Ratio Indices}
\usage{
calc_nri_indices(hyperspecs, indices)
}
\arguments{
\item{hyperspecs}{(\code{list})\cr List of \code{hyperspec} objects}

\item{indices}{(\code{character})\cr Names of indices to calculate}
}
\description{
Calculates Normalized Ratio Indices (NRI) from hyperspectral rasters
}
