% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fun-hyperspectral.R
\name{calc_veg_indices}
\alias{calc_veg_indices}
\title{Calculate vegetation indices}
\usage{
calc_veg_indices(hyperspecs, indices)
}
\arguments{
\item{hyperspecs}{(\code{list})\cr List of \code{hyperspec} objects}

\item{indices}{(\code{character})\cr Names of indices to calculate}
}
\description{
Calculates vegetation indices from hyperspectral rasters
}
