% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{calculate_vi}
\alias{calculate_vi}
\title{calculate_vi}
\usage{
calculate_vi(image)
}
\arguments{
\item{image}{(\code{brick})\cr File name of raster brick/stack.}
}
\value{
calculated VIs (file name)
}
\description{
Calculates the following vegetation indices from the Sentinel data:
\itemize{
\item EVI
\item GDVI2
\item GDVI3
\item GDVI4
\item MNDVI
\item MSR
\item D1
}
}
