% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{create_defoliation_map}
\alias{create_defoliation_map}
\title{write_defoliation_map}
\usage{
create_defoliation_map(data, algorithm, limits, title)
}
\arguments{
\item{data}{(\code{data.frame})\cr Predictions including coordinates}

\item{algorithm}{(\code{character})\cr Name of the algoritm}

\item{limits}{(\code{integer})\cr Y axis limits}
}
\value{
(\code{character})
}
\description{
Predict defoliation for the Basque Country
}
