% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{create_prediction_df}
\alias{create_prediction_df}
\title{create_prediction_df}
\usage{
create_prediction_df(data, model)
}
\arguments{
\item{data}{(\code{sf})\cr Masked vegetation indices.}

\item{model}{Used for ordering the variables names (xgboost)}
}
\value{
(\code{data.frame})
}
\description{
Create the prediction data. Strips variable names and orders them.
}
