% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fun-download.R
\name{download}
\alias{download}
\alias{download_trees}
\alias{download_aoi}
\alias{download_locations}
\alias{download_hyperspectral}
\alias{download_forest_mask}
\title{Download data}
\usage{
download_trees(url)

download_aoi(url)

download_locations(url)

download_hyperspectral(url)

download_forest_mask(url)
}
\arguments{
\item{url}{URL to download from (usually Zenodo)}
}
\description{
\code{download_trees()}: Download in-situ tree data

\code{download_trees()}: Download in-situ tree data

\code{download_locations()}: Download plot location vector data

\code{download_hyperspectral()}: Download hyperspectral data

\code{download_forest_mask()}: Download the forest/non-forest mask.
}
