% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{download_images}
\alias{download_images}
\title{download_images}
\usage{
download_images(records)
}
\arguments{
\item{records}{(\code{character})\cr Names of images to download.}
}
\value{
\code{list} of available records
}
\description{
Downloads the listed Sentinel images from Copernicus Hub.
}
