% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fun-hyperspectral.R
\name{extract_bands_to_plot}
\alias{extract_bands_to_plot}
\title{Extract hyperspectral bands}
\usage{
extract_bands_to_plot(plot_name, buffer, tree_data, hyperspectral_bands)
}
\arguments{
\item{hyperspectral_bands}{(\code{list})\cr List with Raster Bricks of
hyperspectral bands}
}
\description{
Extract indices to trees
}
