% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fun-data-preprocessing.R
\name{extract_coords}
\alias{extract_coords}
\title{extract_coords}
\usage{
extract_coords(data)
}
\arguments{
\item{data}{(\code{list})\cr List containing multiple \code{data.frames}.}
}
\value{
List of \code{data.frames} with X/Y information
}
\description{
Extracts coordinates from a list of \code{data.frames} containing X and Y coordinates
}
