% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fun-hyperspectral.R
\name{extract_indices_to_plot}
\alias{extract_indices_to_plot}
\title{Extract indices}
\usage{
extract_indices_to_plot(plot_name, buffer, tree_data, veg_indices,
  nri_indices)
}
\arguments{
\item{plot_name}{(\code{character})\cr Name of the plot}

\item{buffer}{(\code{integer})\cr Buffer to use for extracing, see \code{\link[raster:extract]{raster::extract()}}.}

\item{tree_data}{(\code{sf})\cr Tree data to extract the indices into.}

\item{veg_indices}{(\code{list})\cr Raster bricks with veg indices}

\item{nri_indices}{(\code{list})\cr Raster bricks with NRI indices}
}
\description{
Extract indices to trees
}
