% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-mlr-paper.R
\name{feature_imp_parallel}
\alias{feature_imp_parallel}
\title{Parallel feature importance wrapper}
\usage{
feature_imp_parallel(task, learner, nmc, measure)
}
\arguments{
\item{task}{mlr task}

\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}

\item{nmc}{Number of Monte Carlo iterations}

\item{measure}{Measure for scoring.}
}
\description{
Calculates feature importance via permutation
}
