% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\docType{package}
\name{FeatureSelection-package}
\alias{FeatureSelection}
\alias{FeatureSelection-package}
\title{FeatureSelection: Research Compendium package for the publication "Placeholder"}
\description{
What the package does (one paragraph)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pat-s/2019-feature-selection}
  \item Report bugs at \url{https://github.com/pat-s/2019-feature-selection/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

}
