% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{file_cp_ts}
\alias{file_cp_ts}
\title{Copy a file to a new location conditionally}
\usage{
file_cp_ts(file_in, dir_out)
}
\arguments{
\item{file_in}{The file to be moved}

\item{dir_out}{The directory to be moved to}
}
\description{
By default the copying is only done if the new file is newer than an
already existing file.
The main goal of this function is to  have one wrapper script which calls
to this function which copy only files that have recently changed.
This avoids lots of uneeded diffs in version control.
}
\examples{
library("fs")
file_touch("test.xy")
file_cp_ts("test.xy", "log/")

# now both files are equal, no copying will be done
file_cp_ts("test.xy", "log/")
}
