% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{get_coordinates}
\alias{get_coordinates}
\title{get_coordinates}
\usage{
get_coordinates(data)
}
\arguments{
\item{data}{(\code{sf})\cr Masked vegetation indices.}
}
\value{
masked VIs (\code{brick})
}
\description{
Extract coordinates from \code{sf} object
}
