% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{get_records}
\alias{get_records}
\title{get_records}
\usage{
get_records(aoi, date, processing_level)
}
\arguments{
\item{aoi}{(\code{sf})\cr Area of interest to download images for}

\item{date}{(\code{character})\cr Date range for which to get the records for}

\item{processing_level}{(\code{character})\cr  Which Sentinel processing level to use}
}
\value{
\code{list} of available records
}
\description{
Splits data into feature sets.
}
