% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-mlr-paper.R
\name{inv_boxcox_rmse}
\alias{inv_boxcox_rmse}
\title{Reverse Boxcox transformation}
\usage{
inv_boxcox_rmse(truth, response)
}
\arguments{
\item{truth}{Truth value}

\item{response}{Response value}
}
\description{
Reverse Boxcox transformation
}
