% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fun-data-preprocessing.R
\name{log_response}
\alias{log_response}
\title{log_response}
\usage{
log_response(data, response)
}
\arguments{
\item{data}{(\code{data.frame})\cr data.frame.}

\item{response}{(\code{character})\cr Name of response.}
}
\value{
\code{data.frame}
}
\description{
Performs a log transformation on the response variable.
}
