% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{mask_vi}
\alias{mask_vi}
\title{mask_vi}
\usage{
mask_vi(image_vi)
}
\arguments{
\item{image_vi}{(\code{character})\cr File name of raster brick/stack.}
}
\value{
masked VIs (\code{brick})
}
\description{
Masks the VIs
}
