% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{mosaic_clouds}
\alias{mosaic_clouds}
\title{mosaic_clouds}
\usage{
mosaic_clouds(records)
}
\arguments{
\item{records}{(\code{list})\cr \code{data.frame} of Sentinel-2 records from \link{get_records}.}

\item{cloud_stack}{cloud stack created by \link{copy_cloud}.}
}
\value{
\code{list} of stacked raster files (Bricks)
}
\description{
Mosaics the cloud cover of a specific date
}
