% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{mosaic_images}
\alias{mosaic_images}
\title{mosaic_images}
\usage{
mosaic_images(records, image_stack)
}
\arguments{
\item{records}{(\code{list})\cr \code{data.frame} of Sentinel-2 records from \link{get_records}.}

\item{image_stack}{Stacked images from \link{stack_bands}.}
}
\value{
\code{list} of mosaiced raster files (Bricks)
}
\description{
Mosaics the images of a specific date
}
