% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fun-data-preprocessing.R
\name{mutate_defol}
\alias{mutate_defol}
\title{mutate_defol}
\usage{
mutate_defol(data)
}
\arguments{
\item{data}{(\code{data.frame})\cr data.frame.}
}
\value{
\code{data.frame}
}
\description{
Mutates the "defoliation" variables so that no absolut zeros occur. These might cause problems when standardizing variables.
}
