% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{predict_defoliation}
\alias{predict_defoliation}
\title{predict_defoliation}
\usage{
predict_defoliation(data, model, coordinates)
}
\arguments{
\item{data}{(\code{data.frame})\cr Prediction data}

\item{model}{Used for ordering the variables names (xgboost)}

\item{coordinates}{Coordinates of the predicted data}
}
\value{
(\code{data.frame})
}
\description{
Predict defoliation for the Basque Country
}
