% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fun-hyperspectral.R
\name{process_hyperspec}
\alias{process_hyperspec}
\alias{process_hyperspec_helper}
\title{process_hyperspec}
\usage{
process_hyperspec(data, id, index, plots, name_out)

process_hyperspec_helper(data, id, index, plots, name_out)
}
\arguments{
\item{data}{(\code{list})\cr Raster bricks}

\item{id}{(\code{character})\cr ID name}

\item{index}{(\code{integer})\cr Internal identifier to check which plot belongs to which image.}

\item{plots}{(\code{sf})\cr sf object containing the plot locations (n = 28)}

\item{name_out}{(\code{character})\cr Name of the plot. Used for naming of the resulting list.}
}
\description{
Preprocessing of hyperspectral data:
\itemize{
\item Crop to plot extent
\item Mask to plot extent
\item Reproject to EPSG 32630
}
}
