% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{ras_to_sf}
\alias{ras_to_sf}
\title{ras_to_sf}
\usage{
ras_to_sf(data)
}
\arguments{
\item{data}{(\code{brick})\cr Masked vegetation indices.}
}
\value{
masked VIs (\code{brick})
}
\description{
Converts masked vegetation indices to \code{sf} objects
}
