% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{scale_defoliation}
\alias{scale_defoliation}
\title{write_defoliation_map}
\usage{
scale_defoliation(data)
}
\arguments{
\item{data}{(\code{data.frame})\cr Defoliation data.frame to scale}
}
\value{
(\code{character})
}
\description{
Scale absolute predicted defoliation to 0 - 100
}
