% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fun-data-preprocessing.R
\name{split_into_feature_sets}
\alias{split_into_feature_sets}
\title{split_into_feature_sets}
\usage{
split_into_feature_sets(data, feature_set)
}
\arguments{
\item{data}{(\code{data.frame})\cr data.frame.}

\item{feature_set}{(\code{character})\cr Name of feature set.}
}
\value{
\code{data.frame}
}
\description{
Splits data into feature sets.
}
