% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{stack_bands}
\alias{stack_bands}
\title{stack_bands}
\usage{
stack_bands(records, image_unzip)
}
\arguments{
\item{records}{(\code{list})\cr \code{data.frame} of Sentinel-2 records from \link{get_records}.}

\item{image_unzip}{File names of unzipped images.}
}
\value{
\code{list} of stacked raster files (Bricks)
}
\description{
Stacks the bands of a specific date
}
