% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fun-data-preprocessing.R
\name{standardize}
\alias{standardize}
\title{standardize}
\usage{
standardize(data, cores)
}
\arguments{
\item{data}{(\code{list})\cr List containing multiple \code{data.frames}.}

\item{cores}{(\code{integer})\cr Number of cores}
}
\value{
List of \code{data.frames} with standardized variables
}
\description{
Standardizes all variables of a list of \code{data.frames} in
parallel
}
\details{
Standardization applies from \code{2:length(names(data))}, so it is
expected that the response is sorted first in the data.
}
