% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-mlr-project.R
\name{train_wrapper}
\alias{train_wrapper}
\title{Create a mlr train wrapper}
\usage{
train_wrapper(learner, tune_object, task)
}
\arguments{
\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}

\item{tune_object}{Optimized hyperparameters}

\item{task}{mlr task}
}
\description{
Creates a mlr train wrapper
}
