% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-mlr-paper.R
\name{tune_ctrl_mbo_30n_70it}
\alias{tune_ctrl_mbo_30n_70it}
\title{mlrMBO 30n 70 iterations tuning setting}
\usage{
tune_ctrl_mbo_30n_70it(param_set)
}
\arguments{
\item{param_set}{mlr Param Set}
}
\description{
mlrMBO 30n 70 iterations tuning setting
}
