% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-mlr-project.R
\name{tune_ctrl_wrapper}
\alias{tune_ctrl_wrapper}
\title{Create a mlr tune control object}
\usage{
tune_ctrl_wrapper(propose.points, iters, n, param_set)
}
\arguments{
\item{propose.points}{See \code{\link[mlrMBO:makeMBOControl]{mlrMBO::makeMBOControl()}}}

\item{iters}{See \code{\link[mlrMBO:setMBOControlTermination]{mlrMBO::setMBOControlTermination()}}}

\item{n}{See \code{\link[mlrMBO:makeMBOControl]{mlrMBO::makeMBOControl()}}}

\item{param_set}{mlr Param Set}
}
\description{
Creates a mlr MBO tune control object
}
