% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-mlr-project.R
\name{tune_wrapper}
\alias{tune_wrapper}
\title{Create a mlr tune wrapper}
\usage{
tune_wrapper(
  learner,
  tune.control,
  show.info,
  par.set,
  resampling,
  measure,
  task
)
}
\arguments{
\item{learner}{Learner from \link[mlr:makeLearner]{mlr::makeLearner}}

\item{tune.control}{See \code{\link[mlr:tuneParams]{mlr::tuneParams()}}}

\item{show.info}{See \code{\link[mlr:tuneParams]{mlr::tuneParams()}}}

\item{par.set}{See \code{\link[mlr:tuneParams]{mlr::tuneParams()}}}

\item{resampling}{mlr resampling}

\item{measure}{Measure for scoring.}

\item{task}{mlr task}
}
\description{
Creates a mlr MBO tune wrapper
}
