% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{unzip_images}
\alias{unzip_images}
\title{download_images}
\usage{
unzip_images(data, pattern)
}
\arguments{
\item{data}{(\code{list})\cr File names of images to unzip.}

\item{pattern}{(\code{charachter})\cr Regex pattern (year)}
}
\value{
\code{list} of unzipped files
}
\description{
Downloads the listed Sentinel images from Copernicus Hub.
}
