% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{write_defoliation_map}
\alias{write_defoliation_map}
\title{write_defoliation_map}
\usage{
write_defoliation_map(data, filename)
}
\arguments{
\item{data}{(\code{data.frame})\cr Predictions including coordinates}

\item{filename}{(\code{character})\cr Filename to write}
}
\value{
(\code{character})
}
\description{
Predict defoliation for the Basque Country
}
