% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{calculate_vi}
\alias{calculate_vi}
\title{calculate_vi}
\usage{
calculate_vi(image)
}
\arguments{
\item{image}{(\code{brick})\cr File name of raster brick/stack.}
}
\value{
calculated VIs (file name)
}
\description{
Calculates vegetation indices from the Sentinel data
}
\details{
Indices:
\itemize{
\item EVI
\item GDVI2
\item GDVI3
\item GDVI4
\item MNDVI
\item MSR
\item D1
}

The following steps are performed:
\enumerate{
\item Take the mosaic and read all raster files (different dates)
\item Calculate the index for all files
\item Stack the single rasters
\item Take the mean from all raster files
\item Write to disk
}
}
