% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-fun-visualization.R
\name{create_prediction_map}
\alias{create_prediction_map}
\title{Create spatial maps from the predicted data}
\usage{
create_prediction_map(
  prediction_raster,
  model_name,
  benchmark_object = NULL,
  resampling = NULL
)
}
\arguments{
\item{prediction_raster}{Predicted raster layer given in a list}

\item{model_name}{Algorithm name}

\item{benchmark_object}{mlr benchmark object containing performance}

\item{resampling}{String with the resampling description that should appear
on the map}
}
\description{
Creates a spatial map plot using a predicted raster layer and
inserts a performance value from a \link[mlr:BenchmarkResult]{mlr::BenchmarkResult}
}
