% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fun-hyperspectral.R
\name{extract_bands_to_plot}
\alias{extract_bands_to_plot}
\title{Extract hyperspectral bands}
\usage{
extract_bands_to_plot(plot_name, buffer, tree_data, hyperspectral_bands)
}
\arguments{
\item{plot_name}{(\code{character})\cr Name of the plot}

\item{buffer}{(\code{integer})\cr Buffer to use for extracing, see \code{\link[raster:extract]{raster::extract()}}.}

\item{tree_data}{(\code{sf})\cr Tree data to extract the indices into.}

\item{hyperspectral_bands}{(\code{list})\cr List with Raster Bricks of hyperspectral bands}
}
\description{
Extract indices to trees
}
