% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-fun-sentinel-processing.R
\name{mask_mosaic}
\alias{mask_mosaic}
\title{mask_mosaic}
\usage{
mask_mosaic(image_mosaic, cloud_mosaic, aoi, forest_mask)
}
\arguments{
\item{image_mosaic}{(\code{list})\cr File names of mosaiced images.}

\item{cloud_mosaic}{\code{list})\cr File names of mosaiced cloud images.}

\item{aoi}{Area to be masked}

\item{forest_mask}{Forest/Non-forest mask}
}
\value{
aoi of stacked raster files (Bricks)
}
\description{
Mask the mosaic
}
