% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90-eda-funs.R
\name{my_pairs}
\alias{my_pairs}
\alias{panel.hist}
\title{EDA functions from A. Brenning}
\usage{
panel.hist(x, col.hist = "cyan", cex.cor = NULL, digits = NULL,
  col = NULL, pch = NULL, cex = NULL, ...)

my_pairs(x, ...)
}
\arguments{
\item{x}{\link{data.frame}}
}
\description{
Correlation plot of a dataset
}
\examples{
my_pairs(iris)
}
